[CDS] Use the source path to get information instead of ucp in URLClassLoader
Summary:
For some URLClassLoader, such as WebappClassLoaderBase which is created
with URL[0], we can't get the Resource through ucp in URLClassLoader.
That causes we can't define the class in EagerAppCDS flow.

To solve the issue, we create a new URL and JarFile with the source path,
and get the information from the JarFile.

Test Plan:
test/jdk/com/alibaba/cds/TestDumpAndLoadClassWithNullURL.java
test/jdk/com/alibaba/cds/TestDumpUnsupportedCheck.java
all cds cases

Reviewers: lei.yul,sanhong.lsh,yunyao.zxl